feat: estruturação de clean architecture e refact de código

+++ src/main/java/com/simplesdental/product/service/CategoryService.java
@@ -1,12 +1,16 @@
 package com.simplesdental.product.service;
 
-import com.simplesdental.product.model.Category;
-import com.simplesdental.product.repository.CategoryRepository;
+import java.util.Optional;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
-import java.util.Optional;
+import com.simplesdental.product.model.Category;
+import com.simplesdental.product.repository.CategoryRepository;
+
+import jakarta.transaction.Transactional;
 
 @Service
 public class CategoryService {
@@ -18,11 +22,15 @@ public class CategoryService {
         this.categoryRepository = categoryRepository;
     }
 
-    public List<Category> findAll() {
-        return this.categoryRepository.findAll();
+    public Page<Category> findAll(Pageable pageable) {
+        return this.categoryRepository.findAll(pageable);
     }
 
     public Optional<Category> findById(Long id) {
+        Optional<Category> category = this.categoryRepository.findById(id);
+        if (category.isPresent()) {
+            return category;
+        }
         return this.categoryRepository.findById(id);
     }
 
@@ -30,7 +38,19 @@ public class CategoryService {
         return this.categoryRepository.save(category);
     }
 
-    public void deleteById(Long id) {
+    @Transactional(rollbackOn = Exception.class)
+    public boolean deleteById(Long id) {
+        if (!this.categoryRepository.existsById(id)) {
+            return false;
+        }
+
         this.categoryRepository.deleteById(id);
+
+        return true;
+    }
+
+    @Transactional(rollbackOn = Exception.class)
+    public void update(Category category) {
+        this.categoryRepository.save(category);
     }
 }
